Improve gslib ascii format support - simple refactor. REDUCED MEMORY USAGE (-1.5 grids now) 	covtable i{x,y,z}node are now shorts

+++ include/dss.h
@@ -146,7 +146,7 @@ typedef struct covtable_lookup_vars_type {
 
     /* auxiliar grids allocated in run_dss */
     float  *covtab;
-    int    *ixnode,
+    short  *ixnode,
            *iynode,
            *iznode;
